backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
autoflake==2.3.1
autopep8==2.3.1
bcrypt==4.2.1
black==24.10.0
click==8.1.7
coverage==7.6.8
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environs==11.2.1
executing==2.1.0
factory_boy==3.3.1
Faker==33.1.0
flake8==7.1.1
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.30.0
isort==5.13.2
jedi==0.19.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
wcwidth==0.2.13
